requirements.txt
django==3.2.19
djangorestframework==3.12.4
mysqlclient==1.4.6
drf_yasg
